refactor(remote): organize hook relay types
