+++ CLI/CMUXCLI+RemoteHookBridge.swift
@@ -10,70 +10,6 @@ extension CMUXCLI {
     private static let remoteHookBridgeMaximumConfigurationBytes = 8 * 1024 * 1024
     private static let remoteHookBridgeMaximumPayloadBytes = 16 * 1024 * 1024
 
-    private struct RemoteHookDescriptor: Codable {
-        let name: String
-        let aliases: [String]
-        let binaryName: String
-        let configDirectory: String
-        let installWhenConfigMissing: Bool
-        let snapshotPaths: [String]
-        let recursivePaths: [String]
-
-        enum CodingKeys: String, CodingKey {
-            case name, aliases
-            case binaryName = "binary_name"
-            case configDirectory = "config_directory"
-            case installWhenConfigMissing = "install_when_config_missing"
-            case snapshotPaths = "snapshot_paths"
-            case recursivePaths = "recursive_paths"
-        }
-    }
-
-    private struct RemoteHookSnapshot: Decodable {
-        let agent: String
-        let action: String
-        let arguments: [String]
-        let entries: [RemoteHookSnapshotEntry]
-    }
-
-    private struct RemoteHookSnapshotEntry: Codable {
-        let path: String
-        let kind: String
-        let contentBase64: String?
-        let mode: UInt16
-
-        enum CodingKeys: String, CodingKey {
-            case path, kind, mode
-            case contentBase64 = "content_base64"
-        }
-    }
-
-    private struct RemoteHookMutation: Encodable {
-        let path: String
-        let delete: Bool
-        let contentBase64: String?
-        let mode: UInt16?
-
-        enum CodingKeys: String, CodingKey {
-            case path, delete, mode
-            case contentBase64 = "content_base64"
-        }
-    }
-
-    private struct RemoteHookPlan: Encodable {
-        let stdoutBase64: String
-        let stderrBase64: String
-        let exitCode: Int32
-        let mutations: [RemoteHookMutation]
-
-        enum CodingKeys: String, CodingKey {
-            case mutations
-            case stdoutBase64 = "stdout_base64"
-            case stderrBase64 = "stderr_base64"
-            case exitCode = "exit_code"
-        }
-    }
-
     func runRemoteHookBridgeCommand(_ arguments: [String]) throws -> Bool {
         guard let command = arguments.first?.lowercased(), command.hasPrefix("__remote-") else {
             return false
