+++ CLI/RemoteHookSnapshot.swift
@@ -1,0 +1,6 @@
+struct RemoteHookSnapshot: Decodable {
+    let agent: String
+    let action: String
+    let arguments: [String]
+    let entries: [RemoteHookSnapshotEntry]
+}
