+++ Sources/RemoteHookInvocation.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct RemoteHookInvocation: Sendable {
+    let arguments: [String]
+    let environment: [String: String]
+    let input: Data
+}
