+++ Sources/RemoteHookInvocationBridgeError.swift
@@ -1,0 +1,4 @@
+struct RemoteHookInvocationBridgeError: Error {
+    let code: String
+    let message: String
+}
