+++ Sources/RemoteHookTransferMetadata.swift
@@ -1,0 +1,4 @@
+struct RemoteHookTransferMetadata: Codable, Sendable {
+    let arguments: [String]
+    let environment: [String: String]
+}
